deployment/app/requirements.txt
a2wsgi==1.10.0
anyio==4.12.1
bcrypt==4.1.3
certifi==2026.2.25
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.110.1
motor==3.3.1
openpyxl==3.1.5
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pymongo==4.5.0
python-dotenv==1.2.1
python-multipart==0.0.22
reportlab==4.4.10
starlette==0.37.2
uvicorn==0.25.0
